pyproject.toml
[project]
name = "nonebot-plugin-kuwo"
version = "0.1.0"
description = "基于 NoneBot2 的酷我音乐插件"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1,<1.0.0",
    "nonebot-adapter-onebot>=2.4.6,<3.0.0",
    "nonebot-plugin-alconna>=0.62.0",
    "nonebot2>=2.5.0,<3.0.0",
    "pydantic>=2.0.0,<3.0.0",
]

[dependency-groups]
dev = [
    "fastapi>=0.115.0,<1.0.0",
    "nonebug>=0.4.4",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "respx>=0.23.0",
    "ruff>=0.15.9",
    "uvicorn>=0.35.0,<1.0.0",
]
